refactor: `Code` 클래스 static 메소드 삭제
